cache: implement snapshots
These snapshots are comparable across time and will determine
invalidated files, allowing Taskrunner to determine whether a task needs
to re-run on boot.

A diff against a previous commit is preferred. For uncommitted (and new)
files, the hash is compared to a hash in the previous snapshot. If there
was no hash before, the file is marked as changed.