*: Refactor extension API
*Subscriptions Before*
Subscriptions would open a channel to listen on and be responsible for
closing the channel.

*Subscriptions After*
taskrunner is responsible for closing channels and indicates that the
process has been terminated through said closure.

*New API*
RunOptions is now Runtime and represents taskrunner's Runtime, which can
be hooked into by extensions.

It now exposes OnStart, OnStop, Subscribe. Subscriptions have been
simplified to only provide a channel. ctx.Done checks can be replaced
with a range over the channel (range ends when the channel is closed).

The mutex for subscriptions can be removed because we are no longer
running the act of subscribing asynchronously. Subscription functions,
however, are still run in goroutines.