config: add ReadDefaultConfig
This new function will be used in place of ReadUserConfig from the
previous config.go. Note that the semantics are slightly different in
that user config used to silently extend workspace config (without
merging tasks). Now that we support an explicit extend attribute,
we no longer do this.

I expect this to mostly have no effect for users unless they messed
with path in the workspace for some reason.