executor: do not run pass if context is done
Stop runPass early if context is canceled. If we're shutting down,
then there's no need to run additional passes.

I originally wanted to do this so that I could overload the
done state for cancellation, but I was worried that might make
canceled tasks kickoff their dependencies. This change would
have clamped down on that. Even without it though, this seems
like the right way to handle shutdown.