executor: add a canceled state
The canceled state is an end state that cannot be recovered from. It is
similar to the error state except invalidations cannot revive the
execution (e.g. user invalidation or keep alive invalidation).

This state is only entered when the context is canceled, e.g.
when the executor is shutting down.

It'd be great to refactor this code so that state interactions
look more like a state machine and are less strewn between files.