task,runner: allow tasks to be marked as Hidden
This will allow us to set a task to Hidden. To start, this will indicate
if a task will be shown in -list or -listall in taskrunner.

This will help avoid devs from using tasks that aren't meant to be
called. For example, in taskrunner devbox/run there are multiple
commands that we want to run in parallel, but could easily be confused
for devbox/run. While if devs truly want to they will still be able to
see all the tasks via -listall.

In the future, this could be used for ordering within the TUI.