task,registry: expose an IsGroup field on task
It may be useful to expose whether a task is a group task or a normal
task to the user.

One possible use case is if there is a custom UI option for taskrunner
(such as the one we've built at Samsara), we want to show whether tasks
are completed or not as they are being executed. However, group tasks
will never complete (even if all the tasks underneath it are done),
which leads to a confusing UX. We would like to hide all group tasks
from our UI, but prior to this commit, there was no way to determine
whether the task was a group task or not.