taskrunner task: update Task type to support options
We would like to support passing options to a task via CLI
(e.g. taskrunner task/name --option)

Update the Task type to include an Options field which allows
devs to specify a list of supported options on the task definition.

Every specified option has the following fields:
* Description - a description of the option's effect
* LongName - the multi-char option name
* ShortName - the single-char option name (specified by rune)
* Default - the default value of the option
* ValueType - the type of value the option specifies (bool/int/string)

Since we only expect options to specify bool/int/string types,
we have provided the following enum for ValueType:
* OptionBoolType
* OptionStringType
* OptionIntType

Eventually, we will pass the options passed to the task as CLI args
into the RunWithOptions fn as a third argument of type map[string]OptionArg.

The OptionArg includes the following fields:
* Value - raw value collected from os.Args
* BoolVal - helper fn that returns bool val for bool options
* IntVal - helper fn that returns int val for int options
* StringVal - helper fn that returns sting val for string options