taskrunner registry: add validation to task Flags field
Ensure that only Run or RunWithFlags is passed, panicing and displaying
a message that lets the dev know that Run is deprecated in favor of
RunWithFlags.

Ensure that flag Long/ShortNames are not dupes/shadows of the task name.

When adding a task to the task registry, validate that all options
specified in the task Flags field:
* Have either a LongName or ShortName specified
* Do not use the reserved "help"/"h" LongName and ShortName
* Include a Description
* Include a ValueType that matches the String/Bool/IntFlagType enum vals

Additionally, ensure that no duplicate LongName or ShortName flags exist
for the same task.