taskrunner executor: pass map of FlagArgs to RunWithFlags
Update executor to also store:
* taskFlagsRegistry - groups all supported flags by task
* taskFlagArgs - groups all passed flags by task

Create default map of flag name to FlagArgs using supported
flags and Default values.

Then parse all flag args passed to the task into the map of FlagArgs,
overriding default values with used passed args/values.
Parsing the passed flags includes verifying that the flags are
supported before converting the list into the map.

Helper fns to cast the original string arg into the relevant types
depending on the TaskFlag's ValueType. If the incorrect
fns are used for the option's ValueType, panic and warn
dev to use the correct helper fn according to the option ValueType.

Eventually pass the map of FlagArgs into the RunWithFlags fn
as the third argument.