taskrunner executor: pass option args to RunWithOptions
Update executor to also store:
* taskOptionsRegistry - groups all supported options by task
* taskOptionArgs - groups all passed options by task

Then parse all option args passed to the task into a map[string]OptionArg.
Parsing the options include verifying that options are supported before
converting the list into the map. Helper fns to cast the original string arg
into the relevant types depending on the Option's ValueType are also included.
If the incorrect fns are used for the option's ValueType, panic and warn
dev to use the correct helper fn according to the option ValueType.

Eventually pass the map[string]OptionArg into the RunWithOptions fn
as the third argument.