taskrunner runner: include invalid task names in desired task set
This change fixes a bug where passing an invalid task name as the
first argument would not trigger an error (e.g. taskrunner invalidtask).

The reason why this was not happening is b/c we were validating whether
the args passed to taskrunner and silently leaving them out of the
map of desired tasks/flags if they were invalid.

This was especially apparent for invalid task names passed as the
first arg since invalid task names passed after a valid task name
were treated as flags (and would error as an invalid flag).

Validation happens within the executor, so let's keep that pattern
and pass invalid task names as part of the desired tasks we pass
into the executor and let validation happen there.