taskrunner runner_internal_test: fix test to verify tasks groups
Due to the type of `expectedGroups` and `taskFlagGroups` (map[string][]string),
we were not verifying whether the task groups themselves were correct. This is
b/c indexing by a `key` that did not exist in either the expected or generated
task groups would result in the default empty array.

Update test to ensure that the expected task groups and the generated task groups
are actually correct.