taskrunner runner: add task validation when grouping tasks/flags
Initially, we solely relied on the executor to validate tasks passed
through the CLI as args.

Unfortunately, this makes it difficult to distinguish between
whether an arg is an argument to a flag being passed into taskrunner
directly (since we allow flag vals to be passed after a space).

To ensure we do not accidentally pass invalid tasks that are meant
to be args to taskrunner flags, validate tasks when grouping
tasks and flags from the CLI args together.