taskrunner goextensions builder: define RunWithFlags
The purpose of `WrapWithGoBuild` is to ensure that that a specified
Go package can build before running the logic passed in `Run`.

`WrapWithGoBuild` does this by cloning the original task, then overwriting
the `Run` fn to first build the specified package  before running any logic
defined in the task's original `Run`.

Notably, `WrapWithGoBuild` defines a `Run` fn on the cloned task regardless
of whether `Run` is defined on the original task.

For tasks that are both:
* wrapped with `WrapWithGoBuild`
* leveraging `RunWithFlags`
this causes a bug to occur where `WrapWithGoBuild`'s cloned task has both
a `Run` fn and a `RunWithFlags` fn.
Both being defined causes the registry to return an error about only
allowing one or the other to be defined.

Update `WrapWithGoBuild` to only define a `Run` fn on the cloned task
if a `Run` fn was included in the original task. Otherwise, default to
defining `RunWithFlags` since `Run` has been deprecated in favor of `RunWithFlags`.