go taskrunner: vendor missing testify/require package
Prior to this change, the test were failing on CI on the master branch
and thus on this branch since it was created from master.

Update vendor directory to include github.com/stretchr/testify/require
which is imported by goextensions/builder_test.go but was missing from
the vendor directory, causing CI tests to fail.